Refactors the rasp interface
