+++ src/poli/core/util/proteins/rasp/__init__.py
@@ -1,0 +1,2 @@
+from .rasp import RaspInterface
+from .rasp import load_cavity_and_downstream_models
