+++ src/poli/core/util/proteins/rasp/inner_rasp/cavity_model.py
