+++ src/poli/core/util/proteins/rasp/inner_rasp/helpers.py
