+++ src/poli/core/util/proteins/rasp/inner_rasp/pdb_parser_scripts/grid.py
