+++ src/poli/core/util/proteins/rasp/inner_rasp/pdb_parser_scripts/parse_pdbs_pred.sh
